Fix missing failure text in Capybara 2.9+
If minitest-capybara is used with Capybara 2.9+ then failures for
certain assertions (like assert_link), will throw this error, rather
than reporting the failure:

NoMethodError: undefined method `failure_message' for Capybara::Helpers:Module

The reason is because the `Capybara::Helpers.failure_message` method was
refactored and disappeared in 2.9:
https://github.com/teamcapybara/capybara/pull/1748

As noted in the pull request, `Capybara::Helpers` module is technically
marked private, so this fixes it by bringing the necessary methods into
minitest-capybara's code base.

I believe this should also solve a potential issue with some of the
`refute_*` helpers returning the incorrect failure message (back when
this was working with Capybara < 2.9). Since `failure_message` was being
called for both assertions and refutations, the message was presumably
the same. This updates things to use a separate
`negative_failure_message` for the refutations (so `refute_link` would
report "expected not to find..." instead of "expected to find...").